refactor: Test함수 Int 제네릭을 Double 형태로 변경

+++ Calculator/CalculatorItemQueueTest/CalculatorItemQueueTest.swift
@@ -9,10 +9,10 @@ import XCTest
 @testable import Calculator
 
 class CalculatorItemQueueTest: XCTestCase {
-    var sut: CalculatorItemQueue<Int>?
+    var sut: CalculatorItemQueue<Double>?
 
     override func setUpWithError() throws {
-        sut = CalculatorItemQueue<Int>()
+        sut = CalculatorItemQueue<Double>()
     }
 
     override func tearDownWithError() throws {
@@ -21,7 +21,7 @@ class CalculatorItemQueueTest: XCTestCase {
     
     func test_초기화후_isEmpty가_true여야한다() {
         //given
-        let sut = CalculatorItemQueue<Int>()
+        let sut = CalculatorItemQueue<Double>()
         
         //when, then
         XCTAssertTrue(sut.isEmpty)
@@ -29,7 +29,7 @@ class CalculatorItemQueueTest: XCTestCase {
     
     func test_초기화후_1을_enqueue하면_isEmpty가_false여야한다() {
         //given
-        var sut = CalculatorItemQueue<Int>()
+        var sut = CalculatorItemQueue<Double>()
         sut.enqueue(1)
         
         //when, then
@@ -48,7 +48,7 @@ class CalculatorItemQueueTest: XCTestCase {
     
     func test_초기화후_1_2_3을_enqueue한후_removeAll을하면_isEmpty가_true여야한다() {
         //given
-        var sut = CalculatorItemQueue<Int>()
+        var sut = CalculatorItemQueue<Double>()
         sut.enqueue(1)
         sut.enqueue(2)
         sut.enqueue(3)
